requirements.txt
PyYAML>=6.0
pathlib>=1.0.1
requests>=2.28.0
python-dotenv>=0.20.0
